# optimize EscapeAscii's Display  and CStr's Debug