@Mark-Simulacrum 2024-01-20
@bors r+

r? @Mark-Simulacrum 

It might be nice to check the code size impact, but my guess is that this isn't really a huge amount of new code. Not sure how much it matters anyway in terms of whether this gets used...